+++ src/ScanPatron/ScanPatron.js
@@ -18,6 +18,7 @@ const ScanPatron = ({ mutator, stripes }) => {
   }, []);
 
   const handleScanPatron = async (values) => {
+    resetDetails();
     const { patronBarcode } = values;
     if (patronBarcode) {
       setLoading(true);
@@ -30,7 +31,8 @@ const ScanPatron = ({ mutator, stripes }) => {
         setScannedPatronDetails(patron[0]);
         setPatronRRAPermission(access?.[0]);
       } else {
-        resetDetails();
+        setScannedPatronDetails(null);
+        setLoading(false);
       }
     } else {
       resetDetails();